requirements.txt
# Cliente WebSocket (car_socket.py, controlador.py)
websocket-client>=1.6.0

# Servidor WebSocket para el simulador (car_simulator.py)
websockets>=12.0
